philoagents-api/external/rootAgentPro/pyproject.toml
[project]
name = "rootAgentPro"
version = "0.1.0"
description = "A Root agent "
authors = [{ name = "Your Name", email = "your@email.com" }]
dependencies = [
    "litellm",
    "python-dotenv",
    "sentence-transformers",
    "google.adk"
    # add other dependencies as needed
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]